fix comments in final review
