+++ src/test/java/buisness/components/NavigationMenuComponent.java
@@ -1,6 +1,6 @@
 package buisness.components;
 
-import buisness.BasePage;
+import buisness.core.BasePage;
 
 import static com.codeborne.selenide.Condition.enabled;
 import static com.codeborne.selenide.Condition.visible;
