+++ src/test/java/tests/Tests.java
@@ -1,9 +1,9 @@
 package tests;
 
-import buisness.DashboardPage;
-import buisness.FiltersPage;
-import buisness.LoginPage;
 import buisness.components.NavigationMenuComponent;
+import buisness.core.DashboardPage;
+import buisness.core.FiltersPage;
+import buisness.service.LoginService;
 import io.qameta.allure.Description;
 import org.junit.jupiter.api.Test;
 
@@ -12,7 +12,7 @@ import static core.utils.constants.MenuNames.FILTERS;
 
 class Tests extends BaseTest {
 
-  private final LoginPage login = new LoginPage();
+  private final LoginService loginService = new LoginService();
   private final DashboardPage dashboardPage = new DashboardPage();
   private final FiltersPage filtersPage = new FiltersPage();
   private final NavigationMenuComponent navigationMenuComponent = new NavigationMenuComponent();
@@ -20,7 +20,8 @@ class Tests extends BaseTest {
   @Test
   @Description("Admin user is able to login and open Filters page")
   void openLoginPageAsAdmin() {
-    login.loginWithUser(ADMINISTRATOR);
+
+    loginService.loginWithUser(ADMINISTRATOR);
     dashboardPage.dashboardPageShouldBeOpened();
     navigationMenuComponent.navigateToMenu(FILTERS);
     filtersPage.filtersPageShouldBeOpened();
